+++ touchkit/src/test/java/org/vaadin/touchkit/itest/oldtests/NavigationManagerView.java
@@ -10,10 +10,8 @@ import org.vaadin.touchkit.ui.VerticalComponentGroup;
 
 import com.vaadin.server.ThemeResource;
 import com.vaadin.ui.Button;
-import com.vaadin.ui.Button.ClickEvent;
-import com.vaadin.ui.Button.ClickListener;
 import com.vaadin.ui.CssLayout;
-import com.vaadin.v7.ui.Label;
+import com.vaadin.ui.Label;
 import com.vaadin.ui.Notification;
 
 public class NavigationManagerView extends AbstractTouchKitIntegrationTest {
@@ -36,13 +34,7 @@ public class NavigationManagerView extends AbstractTouchKitIntegrationTest {
         Button button2 = new Button("Normal Button (with nav theme)");
         button2.setStyleName(StyleNames.BUTTON_NAVIGATION);
         verticalComponentGroup.addComponent(button2);
-        button2.addClickListener(new ClickListener() {
-            
-            @Override
-            public void buttonClick(ClickEvent event) {
-                Notification.show("Hi!");
-            }
-        });
+        button2.addClickListener(event -> Notification.show("Hi!"));
 
         navigationView.setContent(verticalComponentGroup);
